// apps/api/src/modules/ordering/domain/order-status.ts
export enum OrderStatus {
  SUBMITTED = 'SUBMITTED',
  PENDING_MERCHANT = 'PENDING_MERCHANT',
  CHANGE_PROPOSED = 'CHANGE_PROPOSED',
  ACCEPTED = 'ACCEPTED',
  PREPARING = 'PREPARING',
  READY = 'READY',
  FULFILLED = 'FULFILLED',
  COMPLETED = 'COMPLETED',
  CANCELLATION_REQUESTED = 'CANCELLATION_REQUESTED',
  CANCELLED = 'CANCELLED',
  REJECTED = 'REJECTED',
}